LayoutTests/platform/mac/accessibility/parent-of-table-row-is-table-expected.txt

cell 1	cell 2	cell 3
cell 1	cell 2	cell 3
cell 1	cell 2	
1	2
This tests that the parent of an AXRow is not the render table section element (which is unknown), but rather the AXTable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS row.role is 'AXRole: AXRow'
PASS row.parentElement().isEqual(table) is true
PASS successfullyParsed is true

TEST COMPLETE
